streamlit_demo/requirements.txt
streamlit~=1.27.2
streamlit_authenticator
pyyaml
